+++ lib/rihanna/supervisor.ex
@@ -43,7 +43,7 @@ defmodule Rihanna.Supervisor do
 
       {db, config} ->
         db = Keyword.drop(db, [:pool, :pool_size])
-        Supervisor.start_link(__MODULE__, Keyword.merge(config, [db: db]), opts)
+        Supervisor.start_link(__MODULE__, Keyword.merge(config, db: db), opts)
     end
   end